requirements.txt
pandas
openai
python-dotenv
streamlit
groq
seaborn
matplotlib